refactor integration tests to use TestFixture for dummy data creation
